[EDELIVERY-8424] [EDELIVERY-8336] put long instead of int for pMode id [PR #2571]

+++ Domibus-MSH-api/src/main/java/eu/domibus/api/pmode/PModeService.java
@@ -12,7 +12,7 @@ public interface PModeService {
 
     LegConfiguration getLegConfiguration(String messageId) throws PModeException;
 
-    byte[] getPModeFile(int id);
+    byte[] getPModeFile(long id);
 
     PModeArchiveInfo getCurrentPMode();
